add: proposal C for working compatibility group
This proposal is focused on a simple design for metadata about compatibility
within either an existing manifest (or list) or a newly created artifact.
It describes a plugin architecture and namespaced attributes (the metadata)
that are maintained by compatibility interest groups, and a plugin framework
that includes plugins for extracting, checking, and creation, and within
each flexibility for simple compatibility checks (e.g., key/value pair matching)
or more complex graph-based approaches.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>