add proposal D
Proposal D is an extension to Proposal C. Proposal C defines an explicit
example of a compatibility artifact, meaning what a single artifact would
look like paired alongside an image in a registry (in some way) to describe
its compatibility for image selection or similar. Proposal D defines a
compatibility schema that is maintained by a compatibility interest group,
for which the goal is to define the namespace of allowed metadata attributes
and relationships between them. These two proposals are complementary and
would work together to allow for validation and understanding of relationships
between terms, but without adding complexity to the compatibility artifact
(Proposal C) directly.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>